Some rearrangements
