+++ testing/docker-ems-catalogue/ems-test-config.yaml
@@ -14,15 +14,6 @@ event_db:
 
 database_auth: True
 
-#ldap_auth:
-#  ldap_server: 127.0.0.1
-#  ldap_port: 3890
-#  user_dn_format: "uid=%s,cn=users,cn=accounts,dc=jinr,dc=ru"
-#  ldap_username: shift
-#  ldap_password: shift
-#  writer_group_dn: "cn=bmneventwriter,cn=groups,cn=accounts,dc=jinr,dc=ru"
-#  admin_group_dn: "cn=bmneventadmin,cn=groups,cn=accounts,dc=jinr,dc=ru"
-
 title: BM@N Event Metadata System
 
 pages:
